changed architecture: lv ws server seperate from webserver
this means pm2 can start the webserver in cluster mode and fork the lv server
