+++ lib/lvWS.js
@@ -16,15 +16,11 @@ Editor-client:
   3. (ws) C then sends content or structure updates every time the editor sees a change
 */
 
-const express = require("express");
-const router = express.Router();
-const routingUtil = require("../lib/routingUtil");
 const resUtil = require("../lib/resUtil");
 const WebSocketServer = require("ws").Server;
 const tokenProcessor = require("../lib/token");
 const uuidv4 = require("uuid/v4");
 const detailedDiff = require("deep-object-diff").detailedDiff;
-//const deepEqual = require("fast-deep-equal");
 const { logger } = require("../lib/logger");
 const pick = require("object.pick");
 
@@ -88,37 +84,7 @@ require("../lib/database").fullInit.then(collections => {
   access = collections.access;
 });
 
-//expose router
-module.exports = {
-  router: router,
-  giveHttpServer: receiveServer
-};
 
-//POST, GET for fallback but never allows (view) the liveview page
-routingUtil.getAndPost(router, "/:token", function(req, res) {
-  //check for token and code and correct stage (liveview permission mode)
-  routingUtil.fullAuth(req, res,
-    (token, resDoc, codeDoc) =>
-      //send rendered editor page with token set
-      res.render("liveview", {
-        token: token,
-        code: codeDoc.code,
-        accessLevel: codeDoc.level,
-        stage: resDoc.stage
-      }),
-    {
-      permissionMissmatch: (token, resDoc, codeDoc) =>
-        //send no permission page
-        res.render("weakperm", {
-          type: "liveview",
-          token: resDoc.token,
-          stage: resDoc.stage,
-          accessLevel: codeDoc.level
-        }),
-      matchMode: "liveview"
-    }
-  );
-});
 
 //iterates all viewers of the given object
 function forEachObj(obj, callback) {
@@ -692,13 +658,16 @@ function processMessage(clients, tokens, data, ws, accessToken, resDoc) {
   }
 }
 
-//websocket server
-function receiveServer(httpServer) {
+//TODO: logging of ws connection events
+
+//websocket server and expose server accepter
+module.exports = { };
+module.exports = httpServer => {
   //start the WebSocket server
   const wss = new WebSocketServer({
     server: httpServer,
     clientTracking: true,
-    path: "/resolution/liveview/ws"
+    path: "/liveview"
   });
 
   //currently connected viewers and editors
@@ -834,4 +803,4 @@ function receiveServer(httpServer) {
       wsError(ws, "connection error", { err: err });
     });
   });
-}
+};
